Fix TimedStack Call to Process::CLOCK_MONOTONIC
Forcing the global scope here is necessary on some older versions of
Ruby which get confused. This was failing tests on redis-activesupport
on older Rubies, and @Lomey found that the issue stemmed from not having
a global scope here on `Process::CLOCK_MONOTONIC`, therefore the
constant wasn't found. We successfully monkey-patched this in
redis-activesupport to fix our build.